Moves SQS event publishing and dao into sharedutils.
