+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/events/EventPublisher.java
@@ -6,5 +6,11 @@ package gov.cms.bfd.events;
  * that can be converted into JSON.
  */
 public interface EventPublisher {
+  /**
+   * Publish the event object in a manner appropriate to this implementation. For example,
+   * serializing the object to JSON and publishing it as a string to an SQS queue or SNS topic.
+   *
+   * @param event an object to be published
+   */
   void publishEvent(Object event);
 }
